requirements-dev.txt
-e .

pytest==3.6.1
pytest-sanic==0.1.9
pytest-cov==2.5.1
flake8==3.5.0
bumpversion==0.5.3